pyproject.toml
[tool.poetry]
name = "paas-unlucky-dip"
version = "0.1.0"
description = ""
authors = ["Tom Whitwell <tom.whitwell@digital.cabinet-office.gov.uk>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.1"
gunicorn = "^21.2.0"
uvicorn = { extras = ["standard"], version = "^0.23.2" }
boto3 = "^1.28.28"
boto3-stubs = { extras = ["essential"], version = "^1.28.28" }
nh3 = "^0.2.14"
jinja2 = "^3.1.2"
authlib = "^1.2.1"
itsdangerous = "^2.1.2"
httpx = "^0.24.1"
python-ulid = "^1.1.0"
pendulum = "^2.1.2"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
dev = "app.run:dev"
local_prod = "app.run:local_prod"

prod = "app.run:prod"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
